Execute OTP-specific `ex_doc` (from `priv`) (#83)
* Prepare to generate more ex_doc_otp_* (one per major OTP version)

* Make sure multi-OTP -based releases are now possible

The strategy considered is:
1. ask to recreate the _checkouts/rebar3_ex_doc folder
   1.1. this should be done for, e.g., OTP 24 to start
2. don't recreate (further) the _checkouts/rebar3_ex_doc folder
   2.1. continue generation with e.g. OTP 25
3. conditionally download ex_doc from GitHub based on current
   OTP version (e.g. kerl/asdf-activated)

Everything else is pretty much the same, but this commit
also adds a minor change:
- priv/ex_doc_... isn't deleted (it's already ignored and
  is not required to be deleted - maybe a remnant from a
  past iteration?)

* Run CI for a broader OTP/Elixir scope

* Conditionally run ex_doc (escript) from running Erlang/OTP version

We choose based on prior downloaded elements
and the fact we know what we're running with
    erlang:system_info(otp_release).

* Act on CI results: fix GHA workflow

* Act on CI results: respect lib. declaration of min Elixir vsn

* Act on CI results: don't break Dialyzer

"unknown" became a default analysis

* Temporarily taking OTP 27 out of the loop

I want to concentrate on what was working
_before_ 27.0-rc1 came out

* Act on CI results: maybe (?) fix cache path issues

* Act on self-review: fix test case

It doesn't prevent anything from breaking, but it's
"more correct"

* Adapt more recent dev. software versions

* Reverse "Temporarily taking OTP 27 out of the loop"

* Allow for current OTP, OTP-1, OTP-2, for compatibility

We increase our chances of finding an ex_doc we can use for
a specific OTP

* Act on CI results: generate all ex_doc before running CT

We limit concurrency so that each "job" fetches the existing
cache to add to it

* Act on CI results: adapt to upcoming OTP 27 doc. change

* Try to improve on cache save/restore

We want to make sure a job uses prior cache in the same
CI run

* Target, instead of 27 rc, master/main/nightly

Erlang/Elixir/rebar3

This should give us a much more bleeding edge approach, though
it's possible we're increasing the chance of issues

Up to us to decide, later, if we wanna separate this step
into its own workflow which should not be required green
for merge (but kinda seen as a "control" flow)

* Replicate all pre-27 tests as post-27 to adapt input

In pre-27 we assume only the tests that existed;
in post-27 we assume all tests that existed plus the same tests
with new input (as post-27 -ready module)

+++ release.sh
@@ -2,6 +2,7 @@
 
 OTP_RAW_VER=$(erl -eval 'erlang:display(erlang:system_info(otp_release)), halt().'  -noshell)
 OTP_VER=$(echo "${OTP_RAW_VER}" | sed -r 's/\"([0-9]+)\"/\1/g' | sed 's/[^0-9]*//g')
+EXPECTED_OTP=("24" "25" "26" "27")
 
 yes_or_exit() {
     while true; do
@@ -14,9 +15,9 @@ yes_or_exit() {
     done
 }
 
-if [ "${OTP_VER}" != "24" ]; then
+if [ "${OTP_VER}" -lt "24" ]; then
      echo "You are using Erlang/OTP ${OTP_VER}."
-     echo "However, ex_doc must be built using OTP == 24 for compatibility. Giving up..."
+     echo "However, ex_doc must be built using OTP >= 24 for compatibility. Giving up..."
      exit
 fi
 
@@ -27,37 +28,42 @@ if [ "${VERSION}" == "" ]; then
     exit
 fi
 
-rm -f priv/ex_doc_otp_24
+read -r -p "Do you want to recreate _checkouts/rebar3_ex_doc? [yes/no] " yn
+if [ $yn == "yes" ]; then
+    echo "Deleting _checkouts/rebar3_ex_doc..."
+    rm -rf _checkouts/rebar3_ex_doc
 
-mkdir -p  _checkouts
+    mkdir -p  _checkouts
 
-rm -rf _checkouts/rebar3_ex_doc
+    cd _checkouts
 
-cd _checkouts
+    git clone https://github.com/starbelly/rebar3_ex_doc
 
-git clone https://github.com/starbelly/rebar3_ex_doc
-
-cd rebar3_ex_doc
+    cd rebar3_ex_doc
+    mkdir priv
+else
+    cd _checkouts/rebar3_ex_doc
+fi
 
 git checkout "v${VERSION}"
 
 EX_DOC_VER="v$(grep -m1 "@ex_doc_version" mix.exs | awk '{print $2}' | tr -d '"')"
 echo "Fetching ex_doc ${EX_DOC_VER} escript from GitHub ..."
-curl -L -o ex_doc_otp_24 "https://github.com/elixir-lang/ex_doc/releases/download/${EX_DOC_VER}/ex_doc_otp_24"
-curl -L -o ex-doc-otp-24.sha256sum  "https://github.com/elixir-lang/ex_doc/releases/download/${EX_DOC_VER}/ex-doc-otp-24.sha256sum"
+curl -L -o ex_doc_otp_${OTP_VER} "https://github.com/elixir-lang/ex_doc/releases/download/${EX_DOC_VER}/ex_doc_otp_${OTP_VER}"
+curl -L -o ex-doc-otp-${OTP_VER}.sha256sum  "https://github.com/elixir-lang/ex_doc/releases/download/${EX_DOC_VER}/ex-doc-otp-${OTP_VER}.sha256sum"
 
 echo "Validating integrity of ex_doc download..."
 
-sha256sum -c ex-doc-otp-24.sha256sum
+sha256sum -c ex-doc-otp-${OTP_VER}.sha256sum
 
-chmod +x ex_doc_otp_24
-mv ex_doc_otp_24 priv
-mv  ex-doc-otp-24.sha256sum priv
+chmod +x ex_doc_otp_${OTP_VER}
+mv ex_doc_otp_${OTP_VER} priv
+mv  ex-doc-otp-${OTP_VER}.sha256sum priv
 
 echo "Looks good!"
 
-SIZE=$(du -ah priv/ex_doc_otp_24 | awk '{print $1}')
-echo "Size of priv/ex_doc_otp_24: $SIZE"
+SIZE=$(du -ah priv/ex_doc_otp_${OTP_VER} | awk '{print $1}')
+echo "Size of priv/ex_doc_otp_${OTP_VER}: $SIZE"
 
 rebar3 ex_doc
 
@@ -72,5 +78,13 @@ open doc/index.html
 
 echo
 yes_or_exit  "Did everything look ok?"
+
+for OTP in "${EXPECTED_OTP[@]}"
+do
+    if [ ! -f "priv/ex_doc_otp_${OTP}" ]; then
+        echo "Expected file 'priv/ex_doc_otp_${OTP}' does not exist."
+        exit 1
+    fi
+done
 
 rebar3 hex publish
